refactor(ipc): move project:free-memory to defineIpcNamespace codegen
The hand-written maps.ts entry tripped the check:ipc-handwritten ratchet
(new IPC channels must be codegen-generated). Move the handler into a
defineIpcNamespace block so its IpcInvokeMap entry is generated, keeping the
renderer-facing window.electron.project.freeMemory surface unchanged.

- New electron/ipc/handlers/projectFreeMemory.ts namespace (deps-closured op)
- Drop the inline handler from projectCrud/crud.ts and its unused imports
- Register the namespace in handlers.ts; regenerate shared/types/ipc/generated.ts
- Remove the hand-written maps.ts entry (ratchet back to baseline 240)
- Point the handler test at the new registration
